+++ src/aggregators/AveragingAggregator.cpp
@@ -1,9 +1,9 @@
 #include "aggregators/AveragingAggregator.hpp"
 
-int AveragingAggregator::aggregateFitnesses(
-	std::vector<int> apportionedFitnesses
+float AveragingAggregator::aggregateFitnesses(
+	std::vector<float> apportionedFitnesses
 ) {
-	int total = 0;
+	float total = 0;
 	for (unsigned int i = 0; i < apportionedFitnesses.size(); i++)
 		total += apportionedFitnesses[i];
