+++ src/aggregators/BestOfAggregator.cpp
@@ -1,9 +1,9 @@
 #include "aggregators/BestOfAggregator.hpp"
 
-int BestOfAggregator::aggregateFitnesses(
-	std::vector<int> apportionedFitnesses
+float BestOfAggregator::aggregateFitnesses(
+	std::vector<float> apportionedFitnesses
 ) {
-	int bestFitness = 0;
+	float bestFitness = 0;
 	for (unsigned int i = 0; i < apportionedFitnesses.size(); i++)
 		if (apportionedFitnesses[i] > bestFitness)
 			bestFitness = apportionedFitnesses[i];
