+++ src/endconditions/FitnessMatchEnd.cpp
@@ -1,16 +1,16 @@
 #include "endconditions/FitnessMatchEnd.hpp"
 
-FitnessMatchEnd::FitnessMatchEnd(int minimumFitness) {
+FitnessMatchEnd::FitnessMatchEnd(float minimumFitness) {
 	this->minimumFitness = minimumFitness;
 }
 
-bool FitnessMatchEnd::checkSolution(Genome * target, int targetFitness) {
+bool FitnessMatchEnd::checkSolution(Genome * target, float targetFitness) {
 	return targetFitness > this->minimumFitness;
 }
 
 bool FitnessMatchEnd::checkCondition(
 	vector<Genome*> genomes,
-	vector<int> fitnesses,
+	vector<float> fitnesses,
 	unsigned int currentIteration
 ) {
 	for (unsigned int i = 0; i < genomes.size(); i++)
