+++ src/systems/AccelSSGA.cpp
@@ -20,7 +20,7 @@ AccelSSGA::AccelSSGA(
 
 std::vector<Genome*> AccelSSGA::breedMutateSelect(
 	std::vector<Genome*> initialPopulation,
-	std::vector<int> & populationFitnesses,
+	std::vector<float> & populationFitnesses,
 	CrossoverOperation * cross,
 	MutationOperation * mutation,
 	std::vector<ObjectiveFunction*> objectives,
